Check for custom dashboard config
Support both the old and new format for QuickCalls config.

If using the new config format determine the Dashboard loading the
portlet and check for a custom dashboard config. Fallback to the Default
config setting if no custom config.

If there is a dashboard id add hidden input to the form so it gets
passed on when submitted.